+++ toolkit/src/tools/ovtkStimulationSet.cpp
@@ -4,7 +4,7 @@ using namespace OpenViBE;
 
 bool OpenViBEToolkit::Tools::StimulationSet::shift(IStimulationSet& rStimulationSet, const uint64_t ui64TimeShift)
 {
-	const uint64_t count = rStimulationSet.getStimulationCount();
+	const size_t count = rStimulationSet.getStimulationCount();
 	for (uint64_t i = 0; i < count; i++) { rStimulationSet.setStimulationDate(i, rStimulationSet.getStimulationDate(i) + ui64TimeShift); }
 	return true;
 }
@@ -19,7 +19,7 @@ bool OpenViBEToolkit::Tools::StimulationSet::copy(IStimulationSet& rDestinationS
 bool OpenViBEToolkit::Tools::StimulationSet::append(IStimulationSet& rDestinationStimulationSet, const IStimulationSet& rSourceStimulationSet,
 													const uint64_t ui64TimeShift)
 {
-	const uint64_t count = rSourceStimulationSet.getStimulationCount();
+	const size_t count = rSourceStimulationSet.getStimulationCount();
 	for (uint64_t i = 0; i < count; i++)
 	{
 		rDestinationStimulationSet.appendStimulation(rSourceStimulationSet.getStimulationIdentifier(i),
@@ -32,7 +32,7 @@ bool OpenViBEToolkit::Tools::StimulationSet::append(IStimulationSet& rDestinatio
 bool OpenViBEToolkit::Tools::StimulationSet::appendRange(IStimulationSet& rDestinationStimulationSet, const IStimulationSet& rSourceStimulationSet,
 														 const uint64_t ui64SourceStartTime, const uint64_t ui64SourceEndTime, const uint64_t ui64TimeShift)
 {
-	const uint64_t count = rSourceStimulationSet.getStimulationCount();
+	const size_t count = rSourceStimulationSet.getStimulationCount();
 	for (uint64_t i = 0; i < count; i++)
 	{
 		const uint64_t date = rSourceStimulationSet.getStimulationDate(i);
@@ -48,7 +48,7 @@ bool OpenViBEToolkit::Tools::StimulationSet::appendRange(IStimulationSet& rDesti
 
 bool OpenViBEToolkit::Tools::StimulationSet::removeRange(IStimulationSet& rStimulationSet, const uint64_t ui64StartTime, const uint64_t ui64EndTime)
 {
-	for (uint64_t i = 0; i < rStimulationSet.getStimulationCount(); i++)
+	for (size_t i = 0; i < rStimulationSet.getStimulationCount(); i++)
 	{
 		const uint64_t date = rStimulationSet.getStimulationDate(i);
 		if (ui64StartTime <= date && date < ui64EndTime) { rStimulationSet.removeStimulation(i--); }
